internal/graph: moving parition code to its own file and moving methods to funcs
